move pad from _doc_utils.py to _dwt.py and add docstring
